ファイル操作関数を移動
