+++ electron/os.js
@@ -1,0 +1,22 @@
+const { ipcMain, dialog } = require('electron')
+
+let fileOpen = ipcMain.handle('file-open', async (event) => {
+    const { canceled, filePaths } = await dialog.showOpenDialog({
+        filters: [{ name: 'Documents', extensions: ['txt'] }],
+    })
+    if (canceled) return { canceled, data: [] }
+    const data = filePaths.map((filePath) =>
+        fs.readFileSync(filePath, { encoding: 'utf8' })
+    )
+    return { canceled, data }
+  })
+  
+let fileSave = ipcMain.handle('file-save', async (event, data) => {
+    const { canceled, filePath } = await dialog.showSaveDialog({
+        filters: [{ name: 'Documents', extensions: ['txt'] }],
+    })
+    if (canceled) { return }
+    fs.writeFileSync(filePath, data)
+  })
+
+module.exports = { fileOpen, fileSave }
